+++ mqtt/mqtt-bridge/src/persist/waking_state/mod.rs
@@ -70,7 +70,7 @@ mod tests {
         let num_elements = 10 as usize;
         for i in 0..num_elements {
             #[allow(clippy::cast_possible_truncation)]
-            let key = Key { offset: i as u32 };
+            let key = Key { offset: i as u64 };
             let publication = Publication {
                 topic_name: i.to_string(),
                 qos: QoS::ExactlyOnce,
@@ -85,7 +85,7 @@ mod tests {
         let mut elements = state.batch(num_elements).unwrap();
         for count in 0..num_elements {
             #[allow(clippy::cast_possible_truncation)]
-            let num_elements = count as u32;
+            let num_elements = count as u64;
             assert_eq!(elements.pop_front().unwrap().0.offset, num_elements)
         }
     }
@@ -96,7 +96,7 @@ mod tests {
         let num_elements = 10 as usize;
         for i in 0..num_elements {
             #[allow(clippy::cast_possible_truncation)]
-            let key = Key { offset: i as u32 };
+            let key = Key { offset: i as u64 };
             let publication = Publication {
                 topic_name: i.to_string(),
                 qos: QoS::ExactlyOnce,
@@ -123,7 +123,7 @@ mod tests {
         // check that the ordering is maintained
         for count in 2..num_elements {
             #[allow(clippy::cast_possible_truncation)]
-            let num_elements = count as u32;
+            let num_elements = count as u64;
             let extracted_offset = loader.try_next().await.unwrap().unwrap().0.offset;
             assert_eq!(extracted_offset, num_elements)
         }
